Add versions to connector
