+++ diff_sync/mocks/mock_connector.py
@@ -3,6 +3,8 @@ class MockConnector:
         self.text = text
         self.shadow = shadow
         self.backup = backup
+        self.client_version = 0
+        self.server_version = 0
 
     def get_text(self):
         return self.text
@@ -13,6 +15,12 @@ class MockConnector:
     def get_backup(self):
         return self.backup
 
+    def get_client_version(self):
+        return self.client_version
+
+    def get_server_version(self):
+        return self.server_version
+
     def set_text(self, text):
         self.text = text
 
@@ -21,3 +29,9 @@ class MockConnector:
 
     def set_backup(self, text):
         self.backup = text
+
+    def set_client_version(self, version):
+        self.client_version = version
+
+    def set_server_version(self, version):
+        self.server_version = version
