+++ diff_sync/mocks/mock_stack.py
